#23 [CHORE] : adapter 패키지 구현
