+++ app/src/main/java/com/sopt/now/gs/feature/reserve/adapter/ReserveCategoryTopAdapter.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.gs.feature.reserve
+package com.sopt.now.gs.feature.reserve.adapter
 
 import android.content.Context
 import android.util.Log
@@ -13,27 +13,23 @@ import com.sopt.now.gs.R
 import com.sopt.now.gs.data.response.ResponseReserveCategoryDto
 import com.sopt.now.gs.feature.util.PriceFormatter
 
-class ReserveCategoryBottomAdapter(
+class ReserveCategoryTopAdapter(
     val context: Context,
     private val items: List<ResponseReserveCategoryDto.Product>,
     private val onItemClicked: (Int) -> Unit,
 ) : BaseAdapter() {
     override fun getView(position: Int, p1: View?, p2: ViewGroup?): View {
         val view: View =
-            LayoutInflater.from(context).inflate(R.layout.item_reserve_category_bottom, null)
+            LayoutInflater.from(context).inflate(R.layout.item_reserve_category_top, null)
 
         val currentItem = items[position]
 
-        view.findViewById<ImageView>(R.id.iv_reserve_category_bottom_image)
+        view.findViewById<ImageView>(R.id.iv_reserve_category_top_image)
             .load(currentItem.image)
-        view.findViewById<TextView>(R.id.tv_reserve_category_bottom_title).text =
+        view.findViewById<TextView>(R.id.tv_reserve_category_top_menu_title).text =
             currentItem.title
-        view.findViewById<TextView>(R.id.tv_reserve_category_bottom_price).text =
+        view.findViewById<TextView>(R.id.tv_reserve_category_top_price).text =
             PriceFormatter.formatPrice(currentItem.price)
-        view.findViewById<TextView>(R.id.tv_reserve_category_bottom_rate).text =
-            currentItem.starRating.toString()
-        view.findViewById<TextView>(R.id.tv_reserve_category_botom_review).text =
-            context.getString(R.string.reserve_menu_review, currentItem.reviewCount)
 
         view.setOnClickListener {
             onItemClicked(currentItem.productId)
